Conformer correction to free energies and  enantioselectivity option
